+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Migrate from `sklearn+numpy` to `pyarrow` for location querying
 - Remove command from Docker file
 - Improve workflow for Docker
+- Get rid of columns enumeration
 
 ## 0.103.0 - 2025-02-02